Module Safemoney.Predefined

module type Custom = sig ... end

Signature for user-defined currency type. Typically one needs to provide the following:

module ISO4217_AFN : Custom

Predefined module for Afghani

module ISO4217_ALL : Custom

Predefined module for Albanian Lek

module ISO4217_DZD : Custom

Predefined module for Algerian Dinar

module ISO4217_AOA : Custom

Predefined module for Angolan Kwanza

module ISO4217_XCD : Custom

Predefined module for East Caribbean Dollar

module ISO4217_ARS : Custom

Predefined module for Argentine Peso

module ISO4217_AMD : Custom

Predefined module for Armenian Dram

module ISO4217_AWG : Custom

Predefined module for Aruban Florin

module ISO4217_AUD : Custom

Predefined module for Australian Dollar

module ISO4217_AZN : Custom

Predefined module for Azerbaijan Manat

module ISO4217_BSD : Custom

Predefined module for Bahamian Dollar

module ISO4217_BHD : Custom

Predefined module for Bahraini Dinar

module ISO4217_BDT : Custom

Predefined module for Bangladeshi Taka

module ISO4217_BBD : Custom

Predefined module for Barbados Dollar

module ISO4217_BYN : Custom

Predefined module for Belarusian Ruble

module ISO4217_BZD : Custom

Predefined module for Belize Dollar

module ISO4217_XOF : Custom

Predefined module for CFA Franc BCEAO

module ISO4217_BMD : Custom

Predefined module for Bermudian Dollar

module ISO4217_BTN : Custom

Predefined module for Bhutanese Ngultrum

module ISO4217_BOB : Custom

Predefined module for Bolivian Boliviano

module ISO4217_BOV : Custom

Predefined module for Bolivian Mvdol

module ISO4217_BAM : Custom

Predefined module for Bosnia-Herzegovina Convertible Marka

module ISO4217_BWP : Custom

Predefined module for Botswanan Pula

module ISO4217_NOK : Custom

Predefined module for Norwegian Krone

module ISO4217_BRL : Custom

Predefined module for Brazilian Real

module ISO4217_BND : Custom

Predefined module for Brunei Dollar

module ISO4217_BGN : Custom

Predefined module for Bulgarian Lev

module ISO4217_BIF : Custom

Predefined module for Burundi Franc

module ISO4217_CVE : Custom

Predefined module for Cabo Verde Escudo

module ISO4217_KHR : Custom

Predefined module for Cambodian Riel

module ISO4217_XAF : Custom

Predefined module for CFA Franc BEAC

module ISO4217_CAD : Custom

Predefined module for Canadian Dollar

module ISO4217_KYD : Custom

Predefined module for Cayman Islands Dollar

module ISO4217_CLP : Custom

Predefined module for Chilean Peso

module ISO4217_CLF : Custom

Predefined module for Chilean Unidad de Fomento

module ISO4217_CNY : Custom

Predefined module for Chinese Yuan Renminbi

module ISO4217_COP : Custom

Predefined module for Colombian Peso

module ISO4217_COU : Custom

Predefined module for Colombian Unidad de Valor Real

module ISO4217_KMF : Custom

Predefined module for Comorian Franc

module ISO4217_CDF : Custom

Predefined module for Congolese Franc

module ISO4217_CRC : Custom

Predefined module for Costa Rican Colon

module ISO4217_CUP : Custom

Predefined module for Cuban Peso

module ISO4217_CUC : Custom

Predefined module for Cuban Peso Convertible

module ISO4217_ANG : Custom

Predefined module for Netherlands Antillean Guilder

module ISO4217_CZK : Custom

Predefined module for Czech Koruna

module ISO4217_DKK : Custom

Predefined module for Danish Krone

module ISO4217_DJF : Custom

Predefined module for Djibouti Franc

module ISO4217_DOP : Custom

Predefined module for Dominican Peso

module ISO4217_EGP : Custom

Predefined module for Egyptian Pound

module ISO4217_SVC : Custom

Predefined module for El Salvador Colon

module ISO4217_ERN : Custom

Predefined module for Eritrean Nakfa

module ISO4217_SZL : Custom

Predefined module for Swazi Lilangeni

module ISO4217_ETB : Custom

Predefined module for Ethiopian Birr

module ISO4217_EUR : Custom

Predefined module for Euro

module ISO4217_FKP : Custom

Predefined module for Falkland Islands Pound

module ISO4217_FJD : Custom

Predefined module for Fiji Dollar

module ISO4217_XPF : Custom

Predefined module for CFP Franc

module ISO4217_GBP : Custom

Predefined module for Pound Sterling

module ISO4217_GMD : Custom

Predefined module for Gambian Dalasi

module ISO4217_GEL : Custom

Predefined module for Georgian Lari

module ISO4217_GHS : Custom

Predefined module for Ghana Cedi

module ISO4217_GIP : Custom

Predefined module for Gibraltar Pound

module ISO4217_GTQ : Custom

Predefined module for Guatemalan Quetzal

module ISO4217_GNF : Custom

Predefined module for Guinean Franc

module ISO4217_GYD : Custom

Predefined module for Guyana Dollar

module ISO4217_HTG : Custom

Predefined module for Haitian Gourde

module ISO4217_HNL : Custom

Predefined module for Honduran Lempira

module ISO4217_HKD : Custom

Predefined module for Hong Kong Dollar

module ISO4217_HUF : Custom

Predefined module for Hungarian Forint

module ISO4217_ISK : Custom

Predefined module for Iceland Krona

module ISO4217_INR : Custom

Predefined module for Indian Rupee

module ISO4217_IDR : Custom

Predefined module for Indonesian Rupiah

module ISO4217_XDR : Custom

Predefined module for Special Drawing Right

module ISO4217_IRR : Custom

Predefined module for Iranian Rial

module ISO4217_IQD : Custom

Predefined module for Iraqi Dinar

module ISO4217_ILS : Custom

Predefined module for New Israeli Sheqel

module ISO4217_JMD : Custom

Predefined module for Jamaican Dollar

module ISO4217_JPY : Custom

Predefined module for Japenese Yen

module ISO4217_JOD : Custom

Predefined module for Jordanian Dinar

module ISO4217_KZT : Custom

Predefined module for Kazakhstani Tenge

module ISO4217_KES : Custom

Predefined module for Kenyan Shilling

module ISO4217_KPW : Custom

Predefined module for North Korean Won

module ISO4217_KRW : Custom

Predefined module for South Korean Won

module ISO4217_KWD : Custom

Predefined module for Kuwaiti Dinar

module ISO4217_KGS : Custom

Predefined module for Kyrgystani Som

module ISO4217_LAK : Custom

Predefined module for Lao Kip

module ISO4217_LBP : Custom

Predefined module for Lebanese Pound

module ISO4217_LSL : Custom

Predefined module for Lesotho Loti

module ISO4217_LRD : Custom

Predefined module for Liberian Dollar

module ISO4217_LYD : Custom

Predefined module for Libyan Dinar

module ISO4217_CHF : Custom

Predefined module for Swiss Franc

module ISO4217_MOP : Custom

Predefined module for Macanese Pataca

module ISO4217_MKD : Custom

Predefined module for Macedonian Denar

module ISO4217_MGA : Custom

Predefined module for Malagasy Ariary

module ISO4217_MWK : Custom

Predefined module for Malawi Kwacha

module ISO4217_MYR : Custom

Predefined module for Malaysian Ringgit

module ISO4217_MVR : Custom

Predefined module for Maldivian Rufiyaa

module ISO4217_MUR : Custom

Predefined module for Mauritius Rupee

module ISO4217_XUA : Custom

Predefined module for ADB Unit of Account

module ISO4217_MXN : Custom

Predefined module for Mexican Peso

module ISO4217_MXV : Custom

Predefined module for Mexican Unidad de Inversion

module ISO4217_MDL : Custom

Predefined module for Moldovan Leu

module ISO4217_MNT : Custom

Predefined module for Mongolian Tugrik

module ISO4217_MAD : Custom

Predefined module for Moroccan Dirham

module ISO4217_MMK : Custom

Predefined module for Myanmar Kyat

module ISO4217_NAD : Custom

Predefined module for Namibia Dollar

module ISO4217_NPR : Custom

Predefined module for Nepalese Rupee

module ISO4217_NZD : Custom

Predefined module for New Zealand Dollar

module ISO4217_NIO : Custom

Predefined module for Cordoba Oro

module ISO4217_NGN : Custom

Predefined module for Nigerian Naira

module ISO4217_OMR : Custom

Predefined module for Omani Rial

module ISO4217_PKR : Custom

Predefined module for Pakistan Rupee

module ISO4217_PAB : Custom

Predefined module for Panamanian Balboa

module ISO4217_PGK : Custom

Predefined module for Papua New Guinean Kina

module ISO4217_PYG : Custom

Predefined module for Paraguayan Guarani

module ISO4217_PEN : Custom

Predefined module for Peruvian Sol

module ISO4217_PHP : Custom

Predefined module for Philippine Peso

module ISO4217_PLN : Custom

Predefined module for Polish Zloty

module ISO4217_QAR : Custom

Predefined module for Qatari Rial

module ISO4217_RON : Custom

Predefined module for Romanian Leu

module ISO4217_RUB : Custom

Predefined module for Russian Ruble

module ISO4217_RWF : Custom

Predefined module for Rwanda Franc

module ISO4217_SHP : Custom

Predefined module for Saint Helena Pound

module ISO4217_WST : Custom

Predefined module for Samoan Tala

module ISO4217_STN : Custom

Predefined module for Sao Tome and Principe Dobra

module ISO4217_SAR : Custom

Predefined module for Saudi Riyal

module ISO4217_RSD : Custom

Predefined module for Serbian Dinar

module ISO4217_SCR : Custom

Predefined module for Seychelles Rupee

module ISO4217_SLL : Custom

Predefined module for Sierra Leonean Leone

module ISO4217_SLE : Custom

Predefined module for Sierra Leonean Leone

module ISO4217_SGD : Custom

Predefined module for Singapore Dollar

module ISO4217_XSU : Custom

Predefined module for Sucre

module ISO4217_SBD : Custom

Predefined module for Solomon Islands Dollar

module ISO4217_SOS : Custom

Predefined module for Somali Shilling

module ISO4217_ZAR : Custom

Predefined module for South African Rand

module ISO4217_SSP : Custom

Predefined module for South Sudanese Pound

module ISO4217_LKR : Custom

Predefined module for Sri Lanka Rupee

module ISO4217_SDG : Custom

Predefined module for Sudanese Pound

module ISO4217_SRD : Custom

Predefined module for Surinam Dollar

module ISO4217_SEK : Custom

Predefined module for Swedish Krona

module ISO4217_CHE : Custom

Predefined module for WIR Euro

module ISO4217_SYP : Custom

Predefined module for Syrian Pound

module ISO4217_TWD : Custom

Predefined module for New Taiwan Dollar

module ISO4217_TJS : Custom

Predefined module for Tajikstani Somoni

module ISO4217_TZS : Custom

Predefined module for Tanzanian Shilling

module ISO4217_THB : Custom

Predefined module for Thai Baht

module ISO4217_TOP : Custom

Predefined module for Tongan Pa'anga

module ISO4217_TTD : Custom

Predefined module for Trinidad and Tobago Dollar

module ISO4217_TND : Custom

Predefined module for Tunisian Dinar

module ISO4217_TRY : Custom

Predefined module for Turkish Lira

module ISO4217_TMT : Custom

Predefined module for Turkmenistan New Manat

module ISO4217_UGX : Custom

Predefined module for Uganda Shilling

module ISO4217_UAH : Custom

Predefined module for Ukrainian Hryvnia

module ISO4217_AED : Custom

Predefined module for UAE Dirham

module ISO4217_USD : Custom

Predefined module for US Dollar

module ISO4217_USN : Custom

Predefined module for US Dollar (Next day)

module ISO4217_UYU : Custom

Predefined module for Peso Uruguayo

module ISO4217_UYI : Custom

Predefined module for Uruguay Peso en Unidades Indexadas

module ISO4217_UYW : Custom

Predefined module for Unidad Previsional

module ISO4217_UZS : Custom

Predefined module for Uzbekistani Sum

module ISO4217_VUV : Custom

Predefined module for Vanuatu Vatu

module ISO4217_VES : Custom

Predefined module for Bolivar Soberano S

module ISO4217_VED : Custom

Predefined module for Bolivar Soberano D

module ISO4217_VND : Custom

Predefined module for Vietnamese Dong

module ISO4217_YER : Custom

Predefined module for Yemeni Rial

module ISO4217_ZMW : Custom

Predefined module for Zambian Kwacha

module ISO4217_ZWL : Custom

Predefined module for Zimbabwe Dollar

module ISO4217_XBA : Custom

Predefined module for Bond Markets Unit European Composite Unit (EURCO)

module ISO4217_XBB : Custom

Predefined module for Bond Markets Unit European Composite Unit (E.M.U.-6)

module ISO4217_XBC : Custom

Predefined module for Bond Markets Unit European Unit of Account 9 (E.U.A.-9)

module ISO4217_XBD : Custom

Predefined module for Bond Markets Unit European Unit of Account 17 (E.U.A.-17)

module ISO4217_XXX : Custom

Predefined module for Non Currency

module ISO4217_XAU : Custom

Predefined module for Gold

module ISO4217_XPD : Custom

Predefined module for Palladium

module ISO4217_XPT : Custom

Predefined module for Platinum

module ISO4217_XAG : Custom

Predefined module for Silver

module Crypto_BTC : Custom

Predefined module for Bitcoin

module Crypto_ETH : Custom

Predefined module for Ethereum

module Crypto_BNB : Custom

Predefined module for Binance Coin

module Crypto_USDT : Custom

Predefined module for Tether USDT

module Crypto_ADA : Custom

Predefined module for Cardano

module Crypto_DOGE : Custom

Predefined module for Dogecoin

module Crypto_LTC : Custom

Predefined module for Litecoin

module Crypto_XRP : Custom

Predefined module for Ripple

module Crypto_XMR : Custom

Predefined module for Monero

module Crypto_SOL : Custom

Predefined module for Solana

module Crypto_DOT : Custom

Predefined module for Polkadot

module Crypto_SHIB : Custom

Predefined module for Shiba inu

module Crypto_XTZ : Custom

Predefined module for Tezos