Module Safemoney_lib.Qv_intf

Rational Representation

Wrapper type for Zarith Q rational type Q which serves as the underlying type for other currency types that are built based on rational values.

Apart from few helper conversion functions, this signature implement most of what is equivalent to the Num typeclass sensibly.

module type S = sig ... end