Module Safemoney_lib.Zv_intf

Integer Representation

Wrapper type for Zarith Z integer type Z which serves as the underlying type for other currency types that are built based on integer values.

Apart from few helper conversion functions, this signature implement most of what is equivalent to the Num typeclass sensibly.

module type S = sig ... end